Get User Records

This function sends a request to a specified URL, appending necessary details in the URL path, and fetches user records based on the count parameter. If count is more than 100, additional requests are sent out in batches of 100 until the count is fulfilled.

Usage

get_user_records(
  request,
  key,
  repo,
  collection,
  count,
  throttle = c(3000, 300)
)

Arguments

request

The request URL.

key

The API key.

repo

The repository.

collection

The collection.

count

The number of records to fetch. If count is greater than 100, the function will send multiple requests in batches of 100.

throttle

A vector specifying the rate limit as c(requests, seconds). If the number of requests is divisible by requests, the function will pause for seconds seconds. Defaults to c(3000, 300).

Value

A list of responses. Each response is a list representing the JSON response from each request.